refactor(server): gate builtin driver config by platform
