+++ crates/openshell-core/src/config.rs
@@ -1038,8 +1038,6 @@ const fn default_ssh_session_ttl_secs() -> u64 {
 
 #[cfg(test)]
 mod tests {
-    #[cfg(unix)]
-    use super::{is_reachable_unix_socket, is_unix_socket};
     use super::{
         ComputeDriverKind, Config, DEFAULT_SERVICE_ROUTING_DOMAIN, GatewayInterceptorBindingPolicy,
         GatewayInterceptorConfig, GatewayInterceptorFailurePolicy, GatewayJwtConfig,
@@ -1048,6 +1046,8 @@ mod tests {
         docker_host_unix_socket_path, docker_socket_responds, normalize_compute_driver_name,
         podman_socket_candidates_from_env, podman_socket_responds,
     };
+    #[cfg(unix)]
+    use super::{is_reachable_unix_socket, is_unix_socket};
     #[cfg(unix)]
     use std::io::{Read as _, Write as _};
     use std::net::SocketAddr;
